# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Bidding
  description: Bidding Management System
host: 'localhost:4000'
basePath: /
tags: 
- name: Bidding
  description: Bidding Management
schemes: 
- http
- https
consumes:
  - application/json
produces:
  - application/json
paths:
  /:
    get:
      summary: Check Server
      description: To check Server
      operationId: checkServer
      responses:
        200:
          description: OK
  /customers/Register:
    post:
      tags:
      - User
      summary: Add details of user
      description: Enter the details of user
      parameters:
      - name: body
        in: body
        description: Details of user want to register
        required: true
        schema: 
         $ref: '#/definitions/UserRegister'
      responses:
        200:
          description: User added successfully
        209:
          description: Email already registered
        500:
          description: Error while encryption
  /customers/Login:
    post:
      tags:
      - User
      summary: Login the user
      description: Enter the details of user
      parameters:
      - name: body
        in: body
        description: Details of user want to register
        required: true
        schema: 
         $ref: '#/definitions/UserLogin'
      responses:
        200:
          description: OK
        404:
          description: Email not found
        403:
          description: Incorrect Password
        401:
          description: Auth Failed
    x-swagger-router-controller: router
  /customers/User/{userId}:
    patch:
      tags:
      - User
      summary: Update the details of user
      description: To update the details of user
      operationId: upadateUser
      parameters:
      - name: userId
        in: path
        description: ID of user to update
        required: true
        type: string
      - name: body
        in: body
        description: Details of user want to update
        required: true
        schema: 
         $ref: '#/definitions/UserUpdate'
      responses:
        200:
          description: User Updated
        500:
          description: Internal Server Error
        404:
          description: User Not Found
  /customers/User/{Id}:
    delete:
      tags:
      - User
      summary: Delete the user
      description: Delete the details of an individual user
      operationId: deleteUser
      parameters:
      - name: Id
        in: path
        description: ID of user to delete
        required: true
        type: string
      responses:
        200:
          description: User deleted
        500:
          description: Internal Server Error
        404:
          description: User Not Found


  /products/Product:
    post:
      tags:
      - Product
      summary: Add details of product
      description: Enter the details of product
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
      - name: body
        in: body
        description: Details of product user want to add
        required: true
        schema: 
         $ref: '#/definitions/Product'
      - in: formData
        name: product_img
        type: file
        description: The file to upload.
      responses:
        200:
          description: product added successfully
        500:
          description: Internal Server Error
  /products/Products:
    get:
      tags:
      - Product
      summary: Get all the products
      description: To get the details of all product
      operationId: getProduct
      parameters: []
      responses:
        200:
          description: Retrieved successfully
        500:
          description: Internal Server Error
    x-swagger-router-controller: router
  /products/Product/{productId}:
    patch:
      tags:
      - Product
      summary: Update the details of product
      description: To update the details of product
      operationId: upadateProduct
      parameters:
      - name: productId
        in: path
        description: ID of product to update
        required: true
        type: string
      - name: body
        in: body
        description: Details of product want to update
        required: true
        schema: 
         $ref: '#/definitions/ProductUpdate'
      responses:
        200:
          description: Product Updated
        500:
          description: Internal Server Error
        404:
          description: Product Not Found
  /products/Product/{Id}:
    delete:
      tags:
      - Product
      summary: Delete the product
      description: Delete the details of an individual product
      operationId: deleteProduct
      parameters:
      - name: Id
        in: path
        description: ID of product to update
        required: true
        type: string
      responses:
        200:
          description: Product deleted
        500:
          description: Internal Server Error
        404:
          description: Product Not Found
  
  
  /bids/Bid:
    post:
      tags:
      - Bid
      summary: Post bid corresponding a product
      description: Enter the details of regarding bid
      parameters:
      - name: body
        in: body
        description: Details of bid user want to put
        required: true
        schema: 
         $ref: '#/definitions/Bid'
      responses:
        200:
          description: bid posted successfully
        500:
          description: Internal Server Error
  /bids/Bids/{productId}:
    get:
      tags:
      - Bid
      summary: Get all the bids of a product
      description: To get the details of all bids of a product
      operationId: getBid
      parameters:
      - name: productId
        in: path
        description: ID of product for which to get all bid
        required: true
        type: string
      responses:
        200:
          description: Retrieved successfully
        500:
          description: Internal Server Error
    x-swagger-router-controller: router

  /bids/Bid/{Id}:
    get:
      tags:
      - Bid
      summary: Get the maximum bid of a product
      description: To get the maximum of all bids of a product
      operationId: getMaxBid
      parameters:
      - name: Id
        in: path
        description: ID of product for which to get maximum bid
        required: true
        type: string
      responses:
        200:
          description: Retrieved successfully
        500:
          description: Internal Server Error
    x-swagger-router-controller: router
    
  /verificationCode/verification:
    post:
      tags:
      - Verification Code
      summary: Send Verification Number
      description: Enter user id at which Verification Number to be send
      parameters:
      - name: body
        in: body
        description: Details of user id at which Verification Number to be send
        required: true
        schema: 
         $ref: '#/definitions/Random'
      responses:
        200:
          description: Verification Code send successfully
        500:
          description: Internal Server Error
    
      deprecated: false
    
definitions:
  UserRegister:
    type: object
    required:
    - user_email
    - user_pswd
    - user_address
    - user_address2
    - user_city
    - user_stateDetails
    - user_mobile
    properties:
      user_email:
        type: string
      user_pswd:
        type: string
      user_address:
        type: string
      user_address2:
        type: string
      user_city:
        type: string
      user_stateDetails:
        type: string
      user_mobile:
        type: string
  UserLogin:
    type: object
    required:
    - user_email
    - user_pswd
    properties:
      user_email:
        type: string
      user_pswd:
        type: string
  UserUpdate:
    type: object
    required:
      - user_email
      - user_address
      - user_address2
      - user_city
      - user_stateDetails
      - user_mobile
    properties:
      user_email:
        type: string
      user_address:
        type: string
      user_address2:
        type: string
      user_city:
        type: string
      user_stateDetails:
        type: string
      user_mobile:
        type: string
  Product:
    type: object
    required:
      - product_name
      - product_bidding_price
      - product_description
      - product_bidding_EndDate
      - product_userId
      - product_userEmail
    properties:
      product_name:
        type: string
      product_bidding_price:
        type: string
      product_description:
        type: string
      product_bidding_EndDate:
        type: string
      product_userId:
        type: string
      product_userEmail:
        type: string
  ProductUpdate:
    type: object
    required:
      - product_name
      - product_bidding_price
      - product_description
      - product_bidding_EndDate
    properties:
      product_name:
        type: string
      product_bidding_price:
        type: string
      product_description:
        type: string
      product_bidding_EndDate:
        type: string
  Bid:
    type: object
    required:
      - tenderer_email
      - bidder_email
      - bid_price
      - product_id
    properties:
      tenderer_email:
        type: string
      bidder_email:
        type: string
      bid_price:
        type: string
      product_id:
        type: string
  Random:
    type: object
    required:
      - user_id
    properties:
      user_id:
        type: string
      user_email:
        type: string